op-e2e: Move geth init utils to e2eutils
