+++ op-e2e/e2eutils/geth/fakepos.go
@@ -1,4 +1,4 @@
-package op_e2e
+package geth
 
 import (
 	"time"
