+++ op-e2e/e2eutils/geth/geth.go
@@ -1,12 +1,10 @@
-package op_e2e
+package geth
 
 import (
-	"crypto/ecdsa"
 	"fmt"
 	"math/big"
 
 	"github.com/ethereum-optimism/optimism/op-service/clock"
-	"github.com/ethereum/go-ethereum/accounts/keystore"
 	"github.com/ethereum/go-ethereum/cmd/utils"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
@@ -23,9 +21,9 @@ import (
 	_ "github.com/ethereum/go-ethereum/eth/tracers/native"
 )
 
-func initL1Geth(cfg *SystemConfig, genesis *core.Genesis, c clock.Clock, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
+func InitL1(chainID uint64, blockTime uint64, genesis *core.Genesis, c clock.Clock, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
 	ethConfig := &ethconfig.Config{
-		NetworkId: cfg.DeployConfig.L1ChainID,
+		NetworkId: chainID,
 		Genesis:   genesis,
 	}
 	nodeConfig := &node.Config{
@@ -38,7 +36,7 @@ func initL1Geth(cfg *SystemConfig, genesis *core.Genesis, c clock.Clock, opts ..
 		HTTPModules: []string{"debug", "admin", "eth", "txpool", "net", "rpc", "web3", "personal", "engine"},
 	}
 
-	l1Node, l1Eth, err := createGethNode(false, nodeConfig, ethConfig, []*ecdsa.PrivateKey{cfg.Secrets.CliqueSigner}, opts...)
+	l1Node, l1Eth, err := createGethNode(false, nodeConfig, ethConfig, opts...)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -50,7 +48,7 @@ func initL1Geth(cfg *SystemConfig, genesis *core.Genesis, c clock.Clock, opts ..
 		clock:     c,
 		eth:       l1Eth,
 		log:       log.Root(), // geth logger is global anyway. Would be nice to replace with a local logger though.
-		blockTime: cfg.DeployConfig.L1BlockTime,
+		blockTime: blockTime,
 		// for testing purposes we make it really fast, otherwise we don't see it finalize in short tests
 		finalizedDistance: 8,
 		safeDistance:      4,
@@ -77,8 +75,8 @@ func defaultNodeConfig(name string, jwtPath string) *node.Config {
 
 type GethOption func(ethCfg *ethconfig.Config, nodeCfg *node.Config) error
 
-// init a geth node.
-func initL2Geth(name string, l2ChainID *big.Int, genesis *core.Genesis, jwtPath string, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
+// InitL2 inits a L2 geth node.
+func InitL2(name string, l2ChainID *big.Int, genesis *core.Genesis, jwtPath string, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
 	ethConfig := &ethconfig.Config{
 		NetworkId: l2ChainID.Uint64(),
 		Genesis:   genesis,
@@ -93,14 +91,14 @@ func initL2Geth(name string, l2ChainID *big.Int, genesis *core.Genesis, jwtPath 
 		},
 	}
 	nodeConfig := defaultNodeConfig(fmt.Sprintf("l2-geth-%v", name), jwtPath)
-	return createGethNode(true, nodeConfig, ethConfig, nil, opts...)
+	return createGethNode(true, nodeConfig, ethConfig, opts...)
 }
 
 // createGethNode creates an in-memory geth node based on the configuration.
 // The private keys are added to the keystore and are unlocked.
 // If the node is l2, catalyst is enabled.
 // The node should be started and then closed when done.
-func createGethNode(l2 bool, nodeCfg *node.Config, ethCfg *ethconfig.Config, privateKeys []*ecdsa.PrivateKey, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
+func createGethNode(l2 bool, nodeCfg *node.Config, ethCfg *ethconfig.Config, opts ...GethOption) (*node.Node, *eth.Ethereum, error) {
 	for i, opt := range opts {
 		if err := opt(ethCfg, nodeCfg); err != nil {
 			return nil, nil, fmt.Errorf("failed to apply geth option %d: %w", i, err)
@@ -111,28 +109,6 @@ func createGethNode(l2 bool, nodeCfg *node.Config, ethCfg *ethconfig.Config, pri
 	if err != nil {
 		n.Close()
 		return nil, nil, err
-	}
-
-	if !l2 {
-		keydir := n.KeyStoreDir()
-		scryptN := 2
-		scryptP := 1
-		n.AccountManager().AddBackend(keystore.NewKeyStore(keydir, scryptN, scryptP))
-		ks := n.AccountManager().Backends(keystore.KeyStoreType)[0].(*keystore.KeyStore)
-
-		password := "foobar"
-		for _, pk := range privateKeys {
-			act, err := ks.ImportECDSA(pk, password)
-			if err != nil {
-				n.Close()
-				return nil, nil, err
-			}
-			err = ks.Unlock(act, password)
-			if err != nil {
-				n.Close()
-				return nil, nil, err
-			}
-		}
 	}
 
 	backend, err := eth.New(n, ethCfg)
